feat: Add anchored headings to blog post
Anchored headings are to be added to the blog body. Blog post query is modified to return a separate list of anchored headings in the blog body. A table of contents is rendered from the anchored heading list, with anchor links to the headings. Anchored headings are rendered with matching ids for page navigation